Fully updated Calendar to be mudular while loading shifts
