+++ frontend/src/css/app.scss
